third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/edit/switch-panels-while-editing-as-html-expected.txt

Verifies inspector doesn't break when switching panels while editing as HTML. crbug.com/485457


Running: testSetUp

Running: selectNode

Running: switchPanels

Running: switchBackToElements
Is editing: false
